tso controller updates
