+++ src/main/java/zos/shell/controller/factory/type/TsoControllerType.java
@@ -1,4 +1,11 @@
 package zos.shell.controller.factory.type;
 
 public class TsoControllerType {
+
+    public enum Name {
+
+        TSO
+
+    }
+
 }
